Fix type for bc_coeff

+++ args/args.py
@@ -302,7 +302,7 @@ args = {
             'default': 0,
         },
         'bc_coeff': {
-            'type': int,
+            'type': float,
             'default': 0.01,
         },
         'verbose': {
